refactor: Cateogory, CategoryWithIsActive 컴포넌트 분리 후 mainPage 에 적용
- categoryWithIsActive 를 위한 add, delete, get 을 담당하는 query, mutation 를 CategoryWithIsActive 로 이전
- saveButtonToggle 함수 CategoryWithIsActive 로 이전
- FredData 와 백엔드 Favorite 데이터의 동기화를 진행하는 useEffect 를 CategoryWithIsActive 로 이전
- user 가 로그인이면 CategoryWithIsActive 컴포넌트 else 는 Category 컴포넌트를 사용하도록 분기
