+++ src/components/category/Category.module.scss
@@ -1,0 +1,8 @@
+.Category {
+	padding-top: 50px;
+	display: grid;
+	width: 100%;
+	grid-template-columns: 30% 30% 30%;
+	justify-content: space-between;
+	row-gap: 50px;
+}
