+++ src/components/categoryWithIsAcitve/CategoryWithIsActive.module.scss
@@ -1,0 +1,26 @@
+.CategoryWithIsActive {
+	padding-top: 50px;
+	display: grid;
+	width: 100%;
+	grid-template-columns: 30% 30% 30%;
+	justify-content: space-between;
+	row-gap: 50px;
+
+	button {
+		width: 100%;
+		height: 50px;
+		border: none;
+		cursor: pointer;
+		transition: 0.2s;
+
+		&:hover {
+			color: #fff;
+			background: #111;
+		}
+
+		&.on {
+			color: #fff;
+			background: #111;
+		}
+	}
+}
