+++ src/components/categoryWithIsAcitve/CategoryWithIsActive.tsx
@@ -1,0 +1,129 @@
+import clsx from 'clsx';
+import styles from './CategoryWithIsActive.module.scss';
+import { SeriessWithIsActive_Interface, Seriess_Type } from '@/types/fredInterface';
+import IndicatorCard from '../cards/indicatorCard/IndicatorCard';
+import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
+import { addFavorite, deleteFavorite, getFavorite } from '@/backendApi/user';
+import const_queryKey from '@/const/queryKey';
+import { useSelector } from 'react-redux';
+import { useEffect, useState } from 'react';
+
+interface CategoryWithIsActive_Intercae {
+	categoryData: Seriess_Type[];
+	currentPage: number;
+	itemsPerPage: number;
+	categoryId: number;
+}
+
+export default function CategoryWithIsActive({
+	categoryData,
+	currentPage,
+	itemsPerPage,
+	categoryId
+}: CategoryWithIsActive_Intercae) {
+	const user = useSelector((state: Store) => state.user);
+	const queryClient = useQueryClient();
+	const [categoryWithIsActive, setCategoryWithActive] = useState<SeriessWithIsActive_Interface[]>([]);
+
+	// useQuery
+	const { data: favorite, isSuccess: isFavoriteExist } = useQuery({
+		queryKey: [const_queryKey.favorite, categoryId],
+		queryFn: () => getFavorite(user.id, categoryId)
+	});
+
+	const addFavoriteMutation = useMutation({
+		mutationFn: ({ userId, seriesId }: { userId: number; seriesId: string }) => addFavorite(userId, seriesId),
+		onSuccess() {
+			queryClient.invalidateQueries({
+				queryKey: [const_queryKey.favorite]
+			});
+
+			alert('add 성공');
+		},
+		onError(error) {
+			console.error(error);
+		}
+	});
+
+	const deleteFavoriteMutation = useMutation({
+		mutationFn: ({ userId, seriesId }: { userId: number; seriesId: string }) => deleteFavorite(userId, seriesId),
+		onSuccess() {
+			queryClient.invalidateQueries({
+				queryKey: [const_queryKey.favorite]
+			});
+			alert('delete 성공');
+		},
+		onError(error) {
+			console.error(error);
+		}
+	});
+
+	const saveButtonToggle = (userId: number, isActive: boolean, seriesId: string) => {
+		if (isActive) {
+			deleteFavoriteMutation.mutate({ userId, seriesId });
+		} else {
+			addFavoriteMutation.mutate({ userId, seriesId });
+		}
+	};
+
+	useEffect(() => {
+		if (categoryData && isFavoriteExist && favorite) {
+			const categoryWithIsActive = categoryData.map((item: Seriess_Type) => ({
+				...item,
+				isActive: false
+			}));
+
+			favorite.forEach((favoriteIndicator: IndicatorWithIsActive) => {
+				categoryWithIsActive.forEach((categoryIndicator: SeriessWithIsActive_Interface) => {
+					if (favoriteIndicator.seriesId === categoryIndicator.id) categoryIndicator.isActive = true;
+				});
+			});
+
+			setCategoryWithActive(categoryWithIsActive);
+		}
+	}, [currentPage, favorite, categoryData]);
+
+	return (
+		<figure className={clsx(styles.CategoryWithIsActive)}>
+			{categoryWithIsActive
+				.slice(currentPage * itemsPerPage, (currentPage + 1) * itemsPerPage)
+				.map((series: SeriessWithIsActive_Interface, idx: number) => {
+					const { title, id: seriesId, frequency, popularity, observation_start, observation_end, isActive } = series;
+					const notes = series.notes ?? '';
+					return (
+						<IndicatorCard
+							key={idx}
+							title={title}
+							seriesId={seriesId}
+							categoryId={categoryId}
+							frequency={frequency}
+							popularity={popularity}
+							notes={notes}
+							observation_end={observation_end}
+							observation_start={observation_start}
+							className={styles.IndicatorCard}>
+							<div>
+								{notes ? (
+									<p>{notes}</p>
+								) : (
+									<p>This indicator does not have information about the indicator description.</p>
+								)}
+							</div>
+							<button
+								className={clsx(isActive ? styles.on : '')}
+								onClick={() => {
+									saveButtonToggle(user.id, isActive, seriesId);
+									setCategoryWithActive(prevArray => {
+										return prevArray.map((item, mapIndex) => {
+											return mapIndex === idx ? { ...item, isActive: !item.isActive } : item;
+										});
+									});
+								}}>
+								{isActive ? 'remove' : 'save'}
+							</button>
+						</IndicatorCard>
+					);
+				})}
+		</figure>
+	);
+}
